refactor: extract connection to connection.rs (Phase 3)
Move Group D (ConnHandle, Connection, ctx channels + callback + recover, accept
helpers, ConnectionShutdownWaiter, and the Connection h3 trait impls) out of
lib.rs into connection.rs. Connection/ConnectionShutdownWaiter re-exported with
pub use (public path msquic_h3::Connection preserved); internal items via
pub(crate) use (test-only ones cfg(test)-gated). Member visibility raised only to
pub(crate). Siblings compile unedited. Pure move; no behavior/API change.
176 tests / 2 ignored.
